requirements.txt
numpy
opencv-python
pillow
tensorflow-gpu
Flask
